DO NOT COMMIT update formatting.
To avoid too much formatting changes in the series fix the maximum column width
back to 80. After the series lands it'll be possible to format all affected
files and update them in one patch and add that patch to the ignore list.

Hopefully at that time clang-format has better support for concepts.