[libc++][RFC] Guard against libc++ macros in tests.
D117992 removed an libc++ specific macro in our tests. This example lit
tests guards against reintroducing this macro in future commits.

Is there an interest in a script like this? If so I'll look at the other
libc++ macros we want to guard against.

Do we want to guard both `test/std` and `test/libcxx` or only the
former?  (The latter is already libc++ specific, but using the TEST
macros there seems good practice.) Currently the script only tests for
`test/std`.

NOTE the test should fail since we don't depend on D117992.

Differential Revision: https://reviews.llvm.org/D117993