[libc++][chrono] Adds time zone support basics.
This is a first step to add time zone support to the chrono header. The
code is experimental and not ready to be used.

Note that the parsers have no public unit tests. The next patch will
implement sys_info which will make it possible to add these tests and
improve the parser.

Implements parts of:
- P0355 Extending <chrono> to Calendars and Time Zones

Addresses:
- LWG3319 Properly reference specification of IANA time zone database

The patch is still under heavy development and rewriting.